+++ pkg/installer/packages/repository/link_test.go
