+++ pkg/installer/packages/repository/repositories_test.go
