+++ pkg/installer/packages/repository/repository_test.go
